Timers & Basic Events
Added support to schedule a callback after a specified time-duration
(timer). This introduces the generic `Event` class which acts as a
handle to any event produced by the underlying libevent. Right now the
only supported action on the event is cancellation, but the event is
intended to be generic across all events produced by libevent.

This lays the foundation for a scheduler and higher level concepts such
as Tasks or Futures.